moved some lib to utils
